refactor: extract handling of initially disabled menu items
