+++ src/ElectronBackend/main/menu.ts
@@ -25,67 +25,16 @@ import {
 } from './listeners';
 import logger from './logger';
 import {
+  activateMenuItems,
+  INITIALLY_DISABLED_ITEMS_INFO,
+} from './menu/initiallyDisabledMenuItems';
+import {
   getPathOfChromiumNoticeDocument,
   getPathOfNoticeDocument,
 } from './notice-document-helpers';
 import { UserSettings } from './user-settings';
 
 import MenuItemConstructorOptions = Electron.MenuItemConstructorOptions;
-
-const INITIALLY_DISABLED_MENU_ITEMS = [
-  'save',
-  'projectMetadata',
-  'projectStatistics',
-  'followUp',
-  'compactComponentList',
-  'detailedComponentList',
-  'spdxYAML',
-  'spdxJSON',
-  'selectAll',
-  'searchAttributions',
-  'searchSignals',
-  'searchResourcesAll',
-  'searchResourceLinked',
-] as const;
-
-type Item = { label: string; id: string };
-
-const INITIALLY_DISABLED_ITEMS_INFO: Record<
-  (typeof INITIALLY_DISABLED_MENU_ITEMS)[number],
-  Item
-> = {
-  save: { label: 'Save', id: 'save' },
-  followUp: { label: 'Follow-Up', id: 'follow-up' },
-  compactComponentList: {
-    label: 'Compact component list',
-    id: 'compact-list',
-  },
-  detailedComponentList: {
-    label: 'Detailed component list',
-    id: 'detailed-list',
-  },
-  spdxYAML: { label: 'SPDX (yaml)', id: 'spdx-yaml' },
-  spdxJSON: { label: 'SPDX (json)', id: 'spdx-json' },
-  projectMetadata: { label: 'Project Metadata', id: 'project-metadata' },
-  projectStatistics: {
-    label: 'Project Statistics',
-    id: 'project-statistics',
-  },
-  selectAll: { label: 'Select All', id: 'select-all' },
-  searchAttributions: {
-    label: 'Search Attributions',
-    id: 'search-attributions',
-  },
-  searchSignals: { label: 'Search Signals', id: 'search-signals' },
-  searchResourcesAll: {
-    label: 'Search All Resources',
-    id: 'search-resources-all',
-  },
-  searchResourceLinked: {
-    label: 'Search Linked Resources',
-    id: 'search-resources-linked',
-  },
-};
 
 export const importFileFormats: Array<FileFormatInfo> = [
   {
@@ -576,16 +525,4 @@ export async function createMenu(mainWindow: BrowserWindow): Promise<Menu> {
     getAboutMenu(),
     getHelpMenu(webContents),
   ]);
-}
-
-export function activateMenuItems(): void {
-  const menu = Menu.getApplicationMenu();
-  INITIALLY_DISABLED_MENU_ITEMS.forEach((key) => {
-    const menuItem = menu?.getMenuItemById(
-      INITIALLY_DISABLED_ITEMS_INFO[key].id,
-    );
-    if (menuItem) {
-      menuItem.enabled = true;
-    }
-  });
 }
